requirements.txt
fastapi==0.115.12
langchain==0.3.25
langchain_community==0.3.24
langchain_huggingface==0.2.0
python-dotenv==1.1.0
Requests==2.32.3
streamlit==1.45.1
